cambio de nombre a clase
